@@ -8,7 +8,8 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "serve": "concurrently \"ng serve\" \"server/node_modules/.bin/ts-node-dev server/server.ts\""
   },
   "private": true,
   "dependencies": {
@@ -21,11 +22,14 @@
     "@angular/platform-browser": "^5.2.0",
     "@angular/platform-browser-dynamic": "^5.2.0",
     "@angular/router": "^5.2.0",
+    "@types/typescript": "^2.0.0",
     "angular-spotify": "^1.5.1",
     "angular2-jwt": "^0.2.3",
     "angular2-spotify": "^1.3.5",
+    "body-parser": "^1.18.3",
     "bootstrap": "^4.1.3",
     "core-js": "^2.4.1",
+    "express": "^4.16.3",
     "rxjs": "^5.5.6",
     "simple-oauth2": "^2.2.1",
     "spotify-oauth-workflow": "0.0.1",
@@ -35,7 +39,7 @@
     "@angular/cli": "1.6.5",
     "@angular/compiler-cli": "^5.2.0",
     "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@types/jasmine": "~2.5.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
     "codelyzer": "^4.0.1",
@@ -49,6 +53,6 @@
     "protractor": "^5.4.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "^3.0.1"
   }
 }
